requirements.txt
Keras
tensorflow-gpu
scipy
matplotlib
numpy
tqdm
opencv-python
baker
json
cytoolz
lxml
pandas
